Declare Resource class in commonMain instead of using expect/actual
