+++ library/binary-core/src/commonMain/kotlin/io/matthewnelson/kmp/tor/binary/core/PlatformResource.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+
+package io.matthewnelson.kmp.tor.binary.core
+
+/**
+ * Platform specific arguments for Jvm/Js/Native resources
+ * */
+@InternalKmpTorBinaryApi
+public expect class PlatformResource {
+
+    @KmpTorBinaryCoreDsl
+    @InternalKmpTorBinaryApi
+    public class Builder internal constructor() {
+
+        internal fun build(): PlatformResource?
+    }
+
+    override fun equals(other: Any?): Boolean
+    override fun hashCode(): Int
+    override fun toString(): String
+}
