+++ library/binary/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/binary/internal/-JvmPlatform.kt
@@ -28,14 +28,20 @@ internal actual val RESOURCE_CONFIG: Resource.Config by lazy {
 
         resource(ALIAS_GEOIP) {
             isExecutable = false
-            resourceClass = clazz
-            resourcePath = PATH_RESOURCE_GEOIP
+
+            platform {
+                resourceClass = clazz
+                resourcePath = PATH_RESOURCE_GEOIP
+            }
         }
 
         resource(ALIAS_GEOIP6) {
             isExecutable = false
-            resourceClass = clazz
-            resourcePath = PATH_RESOURCE_GEOIP6
+
+            platform {
+                resourceClass = clazz
+                resourcePath = PATH_RESOURCE_GEOIP6
+            }
         }
 
         val host = OSInfo.INSTANCE.osHost
@@ -56,8 +62,11 @@ internal actual val RESOURCE_CONFIG: Resource.Config by lazy {
 
         resource(ALIAS_TOR) {
             isExecutable = true
-            resourceClass = clazz
-            resourcePath = torResourcePath
+
+            platform {
+                resourceClass = clazz
+                resourcePath = torResourcePath
+            }
         }
     }
 }
